Test Fixes

+++ src/Traits/IsLockable.php
@@ -22,8 +22,7 @@ trait IsLockable
         if (config('laravel-lockable.get_locked_on_retrieve', true)) {
             static::retrieved(function (Model $model) {
                 if (! empty($model->lockable)) {
-                    if (!empty($model->lockable->user))
-                    {
+                    if (!empty($model->lockable->user)) {
                         $model->lockHolderName = $model->lockable->user->name;
                     }
                     $model->lockWatcherUsers = $model->lockable->lockWatcherUsers;
@@ -41,7 +40,11 @@ trait IsLockable
                     return true;
                 }
 
-                if (! empty($model->lockable) && $model->lockable->user_id == Auth::id()) {
+                if (!empty($model->lockable) && $model->lockable->user_id == Auth::id()) {
+                    return true;
+                }
+
+                if (empty($model->lockable)) {
                     return true;
                 }
 
@@ -50,8 +53,10 @@ trait IsLockable
         }
 
         static::updated(function (Model $model) {
-            if ($model->lockable->user_id == Auth::id()) {
-                $lockables = $model->lockable->first()->delete();
+            if (!empty($model->lockable)) {
+                if ($model->lockable->user_id == Auth::id()) {
+                    $lockables = $model->lockable->first()->delete();
+                }
             }
         });
     }
@@ -114,7 +119,10 @@ trait IsLockable
         if (! $this->acquiringLock) {
             $this->acquiringLock = true;
         }
-        $lockables = $this->lockable->first()->delete();
+        if (!empty($this->lockable)) {
+            $lockables = $this->lockable->first()->delete();
+        }
+
 
         return true;
     }
